refactor: 토너먼트 라우팅 [id] 변경 및 _common 콜로케이션 적용
