+++ apps/web/src/app/tournament/[id]/_common/_components/TournamentClient.tsx
