+++ apps/web/src/app/tournament/[id]/_common/_consts/rounds.ts
