+++ apps/web/src/app/tournament/[id]/_common/_hooks/useTournament.ts
