+++ apps/web/src/app/tournament/[id]/_common/_types/tournament.ts
