+++ apps/web/src/app/tournament/[id]/_common/_utils/pairItems.ts
